Fix bugs

+++ MonEx/BrowseOffersViewController.swift
@@ -144,15 +144,15 @@ class BrowseOffersViewController: UIViewController {
 extension BrowseOffersViewController: UITableViewDataSource, UITableViewDelegate{
     
     
-    
-    
-    
-    
-    
     func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCellEditingStyle, forRowAt indexPath: IndexPath) {
         if editingStyle == UITableViewCellEditingStyle.delete{
             if case .results( _) = getOffers.currentStatus{
                 let offer = arrayOfOffers[indexPath.row]
+                if offer.offerStatus.rawValue != Constants.offerStatus.nonActive{
+                    canNotDelete()
+                    return 
+                }
+                
                 deleteAllTheInfo(bidId: offer.bidId!)
                 arrayOfOffers.remove(at: indexPath.row)
                 tableView.deleteRows(at: [indexPath], with: .fade)
@@ -160,7 +160,7 @@ extension BrowseOffersViewController: UITableViewDataSource, UITableViewDelegate
         }
     }
     
-   
+    
     
     
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
@@ -258,28 +258,14 @@ extension BrowseOffersViewController: UITableViewDataSource, UITableViewDelegate
         tableView.deselectRow(at: indexPath, animated: true)
     }
     
-    //we delete the bid and all the info pretending to it
-    func deleteAllTheInfo(bidId: String){
-        
-        var bidsDictionary = appUser.bidIds
-        let index = bidsDictionary.index(of: bidId)
-        if let indexOfBid = index{
-            bidsDictionary.remove(at: indexOfBid)
-            appUser.bidIds = bidsDictionary
-        }
-        
-        
-        let pathForBidStatus = "/bidIdStatus/\(bidId)" // set to Null
-        let pathForBidLocation = "/offerBidsLocation/\(bidId)" // set to Null
-        let pathToMyBids = "/Users/\(self.appUser.firebaseId)/Bid/\(bidId)" // set to null
-        let pathBid = "/\(bidId)"
-        rootReference.updateChildValues([pathForBidStatus: NSNull(), pathForBidLocation: NSNull(), pathToMyBids: NSNull(), pathBid: NSNull()])
-
+    
+    //we use this function to erase the data when there is no response.
+    func deleteInfo(bidId:String){
         rootReference.child("bidIdStatus/\(bidId)").observeSingleEvent(of: .value, with:{ (snapshot) in
             guard let dictionary = snapshot.value as? [String: Any] else{
                 return
             }
-            
+        
             guard let authorOfTheBid = dictionary[Constants.publicBidInfo.authorOfTheBid] as? String else{
                 return
             }
@@ -288,54 +274,19 @@ extension BrowseOffersViewController: UITableViewDataSource, UITableViewDelegate
                 return
             }
             
-            self.appUser.getOtherOffer(bidId: bidId){ otherOffer in
-                
-                let pathForCounterOffer = "/counterOffer/\(authorOfTheBid)/\(bidId)"//set to null
-                //set to Null
-                let pathForCounterOfferOther = "/counterOffer/\(otherUser)/\(bidId)"
-                
-                if let otherOffer = otherOffer{
-                    let pathForTranspose = "/transposeOfacceptedOffer/\(otherOffer.firebaseIdOther!)/\(bidId)"// set to Null
-                    self.rootReference.updateChildValues([pathForBidStatus: NSNull(), pathForBidLocation: NSNull(), pathForTranspose: NSNull(), pathToMyBids: NSNull()])
-                    self.rootReference.setValue([pathForCounterOffer:NSNull()])
-                    self.rootReference.setValue([pathForCounterOfferOther: NSNull()])
-                }else{
-                    self.rootReference.setValue([pathForCounterOffer:NSNull()])
-                    self.rootReference.setValue([pathForCounterOfferOther: NSNull()])
-                }
-               
+            
+            //if the author of the bid is not the user the we erase the path to my bids otherwise is already updated
+            if authorOfTheBid != self.appUser.firebaseId {
+                let pathToMyBids = "/Users/\(self.appUser.firebaseId)/Bid/\(bidId)" //set to null
+                self.rootReference.updateChildValues([pathToMyBids: NSNull()])
             }
             
-        })
-    }
-
-    //we use this function to errase the data when there is no response.
-    func deleteInfo(bidId:String){
-        rootReference.child("bidIdStatus/\(bidId)").observeSingleEvent(of: .value, with:{ (snapshot) in
-            guard let dictionary = snapshot.value as? [String: Any] else{
-                return
-            }
-            guard let lastOneToWrite = dictionary[Constants.publicBidInfo.lastOneToWrite] as? String else{
-                return
-            }
+            let pathForCounterOffer = "/counterOffer/\(authorOfTheBid)/\(bidId)"//set to null
+            //set to Null
+            let pathForCounterOfferOther = "/counterOffer/\(otherUser)/\(bidId)"//set to null
+            self.rootReference.updateChildValues([pathForCounterOffer: NSNull()])
+            self.rootReference.updateChildValues([pathForCounterOfferOther:NSNull()])
             
-            // if the last one to write was the user then everything that was created for the bid should be erased
-            if lastOneToWrite == self.appUser.firebaseId{
-                
-                self.appUser.getOtherOffer(bidId: (bidId)){ otherOffer in
-                    
-                    guard let otherOffer = otherOffer else{
-                        return
-                    }
-                    
-                    let pathForBidStatus = "/bidIdStatus/\(bidId)" // set to Null
-                    let pathForTranspose = "/transposeOfacceptedOffer/\(otherOffer.firebaseIdOther!)/\(bidId)"//set to null
-                    let pathForBidLocation = "/offerBidsLocation/\(bidId)/lastOfferInBid/\(Constants.offer.offerStatus)" //update to non active
-                    let pathToMyBids = "/Users/\(self.appUser.firebaseId)/Bid/\(bidId)" //set to null
-                    
-                    self.rootReference.updateChildValues([pathForBidStatus: NSNull(), pathForBidLocation: Constants.offerStatus.nonActive, pathForTranspose: NSNull(), pathToMyBids: NSNull()])
-                }
-            }
         })
     }
 
@@ -485,6 +436,8 @@ extension BrowseOffersViewController: UITableViewDataSource, UITableViewDelegate
         }
     }
     
+    
+    
     func showExpiredAlert(){
         let alert = UIAlertController(title: NSLocalizedString("The request has expired", comment: "The request has expired"), message: "The request that have not been approved expire after 5 min", preferredStyle: .alert)
         let action = UIAlertAction(title: "OK", style: .default, handler:{ (alert) in
@@ -494,4 +447,65 @@ extension BrowseOffersViewController: UITableViewDataSource, UITableViewDelegate
         present(alert, animated: true, completion: nil)
     }
     
+    func canNotDelete(){
+        let alert = UIAlertController(title: NSLocalizedString("Can not delete", comment: "Can not delete"), message: "To Delete this offer you need to reject it or terminate it", preferredStyle: .alert)
+        let action = UIAlertAction(title: "OK", style: .default, handler:{ (alert) in
+            self.getTheOffers()
+        })
+        alert.addAction(action)
+        present(alert, animated: true, completion: nil)
+    }
+    
+    
+    //we delete the bid and all the info pretending to it
+    func deleteAllTheInfo(bidId: String){
+        
+        var bidsDictionary = appUser.bidIds
+        let index = bidsDictionary.index(of: bidId)
+        if let indexOfBid = index{
+            bidsDictionary.remove(at: indexOfBid)
+            appUser.bidIds = bidsDictionary
+        }
+        
+        
+        let pathForBidStatus = "/bidIdStatus/\(bidId)" // set to Null
+        let pathForBidLocation = "/offerBidsLocation/\(bidId)" // set to Null
+        let pathToMyBids = "/Users/\(self.appUser.firebaseId)/Bid/\(bidId)" // set to null
+        let pathBid = "/\(bidId)"
+        rootReference.updateChildValues([pathForBidStatus: NSNull(), pathForBidLocation: NSNull(), pathToMyBids: NSNull(), pathBid: NSNull()])
+        
+        rootReference.child("bidIdStatus/\(bidId)").observeSingleEvent(of: .value, with:{ (snapshot) in
+            guard let dictionary = snapshot.value as? [String: Any] else{
+                return
+            }
+            
+            guard let authorOfTheBid = dictionary[Constants.publicBidInfo.authorOfTheBid] as? String else{
+                return
+            }
+            
+            guard let otherUser = dictionary[Constants.publicBidInfo.otherUser] as? String else{
+                return
+            }
+            
+            self.appUser.getOtherOffer(bidId: bidId){ otherOffer in
+                
+                let pathForCounterOffer = "/counterOffer/\(authorOfTheBid)/\(bidId)"//set to null
+                //set to Null
+                let pathForCounterOfferOther = "/counterOffer/\(otherUser)/\(bidId)"
+                
+                if let otherOffer = otherOffer{
+                    let pathForTranspose = "/transposeOfacceptedOffer/\(otherOffer.firebaseIdOther!)/\(bidId)"// set to Null
+                    self.rootReference.updateChildValues([pathForBidStatus: NSNull(), pathForBidLocation: NSNull(), pathForTranspose: NSNull(), pathToMyBids: NSNull()])
+                    self.rootReference.setValue([pathForCounterOffer:NSNull()])
+                    self.rootReference.setValue([pathForCounterOfferOther: NSNull()])
+                }else{
+                    self.rootReference.setValue([pathForCounterOffer:NSNull()])
+                    self.rootReference.setValue([pathForCounterOfferOther: NSNull()])
+                }
+                
+            }
+            
+        })
+    }
+
 }
